server-side/requirements.txt
black==23.7.0
blinker==1.6.2
click==8.1.6
colorlog==6.7.0
Flask==2.3.2
importlib-metadata==6.8.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2
